+++ examples/resources/pingfederate_oauth_ciba_server_policy_settings/import.sh
@@ -1,0 +1,2 @@
+# "oauthCibaServerPolicySettingsId" should be "id"
+terraform import pingfederate_oauth_ciba_server_policy_settings.myOauthCibaServerPolicySettings oauthCibaServerPolicySettingsId