+++ internal/resource/common/resourcelink/resource_link_schema.go
@@ -2,8 +2,6 @@ package resourcelink
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 )
 
 func ToSchema() map[string]schema.Attribute {
@@ -16,9 +14,6 @@ func ToSchema() map[string]schema.Attribute {
 			Description: "A read-only URL that references the resource. If the resource is not currently URL-accessible, this property will be null.",
 			Computed:    true,
 			Optional:    false,
-			PlanModifiers: []planmodifier.String{
-				stringplanmodifier.UseStateForUnknown(),
-			},
 		},
 	}
 }